+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 - [#2090] Support for Mochicrypt packed binarydata tags - loading SWF as subtree
 - [#2079] Replace DefineSprite with GIF, Bulk import sprites from GIFs, also from commandline
 - [#116] Show invalid utf-8 bytes in Strings as `{invalid_utf8:xxx}`
+- [#2097] Commandline command `-header` to modify SWF header values
 
 ### Fixed
 - Close action on SWF inside DefineBinaryData
@@ -3145,12 +3146,13 @@ Major version of SWF to XML export changed to 2.
 [alpha 7]: https://github.com/jindrapetrik/jpexs-decompiler/releases/tag/alpha7
 [#2090]: https://www.free-decompiler.com/flash/issues/2090
 [#2079]: https://www.free-decompiler.com/flash/issues/2079
+[#116]: https://www.free-decompiler.com/flash/issues/116
+[#2097]: https://www.free-decompiler.com/flash/issues/2097
 [#2093]: https://www.free-decompiler.com/flash/issues/2093
 [#1678]: https://www.free-decompiler.com/flash/issues/1678
 [#2094]: https://www.free-decompiler.com/flash/issues/2094
 [#2095]: https://www.free-decompiler.com/flash/issues/2095
 [#223]: https://www.free-decompiler.com/flash/issues/223
-[#116]: https://www.free-decompiler.com/flash/issues/116
 [#1449]: https://www.free-decompiler.com/flash/issues/1449
 [#2070]: https://www.free-decompiler.com/flash/issues/2070
 [#2073]: https://www.free-decompiler.com/flash/issues/2073